requirements.txt
absl-py==2.3.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.4.0
cffi==2.0.0
click==8.1.8
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
flatbuffers==25.9.23
fonttools==4.60.1
greenlet==3.2.0
h11==0.14.0
idna==3.10
jax==0.7.2
jaxlib==0.7.2
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.14
ml_dtypes==0.5.3
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
postgres==4.0
protobuf==4.25.8
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
